+++ src/Tbx/Console/TagShow.php
@@ -19,8 +19,8 @@ class TagShow extends Iface
     protected function configure()
     {
         $this->setName('tagShow')
-            ->setAliases(array('ts'))
-            ->addOption('nextTag', 't', InputOption::VALUE_NONE, 'Display next release tag.')
+            ->setAliases(array('tags', 'ts'))
+            ->addOption('nextTag', 't', InputOption::VALUE_NONE, 'Display next release tag if this project is to be released.')
             ->addOption('noLibs', 'X', InputOption::VALUE_NONE, 'Do not update the ttek libs.')
             ->addOption('dryRun', 'D', InputOption::VALUE_NONE, 'Test how the update would run without uploading changes.')
             ->setDescription("Run from the root of a ttek project.");
@@ -38,42 +38,20 @@ class TagShow extends Iface
 
         if (!\Tbx\Git::isGit($this->getCwd()))
             throw new \Tk\Exception('Not a GIT repository: ' . $this->getCwd());
+
+        $sformat = '<info>%-25s</info> <comment>%-12s %-12s</comment>';
         $vcs = \Tbx\Git::create($this->getCwd(), $input->getOption('dryRun'));
         $vcs->setInputOutput($input, $output);
-        $this->writeInfo(ucwords($this->getName()) . ': ' . basename($vcs->getPath()));
+        $projName = basename($vcs->getPath());
 
-        $tag = $vcs->getCurrentTag();
-        if ($tag)
-            $this->writeComment($tag);
-        if ($input->getOption('nextTag')) {
-            $nextTag = $vcs->lookupNewTag($input->getOptions());
-            $this->writeBlue($nextTag);
-        }
-
-        if ($input->getOption('noLibs') || !count($this->getVendorPaths())) return;
-        foreach ($this->getVendorPaths() as $vPath) {
-            $libPath = rtrim($vcs->getPath(), '/') . $vPath;
-            if (is_dir($libPath)) {      // If vendor path exists
-                foreach (new \DirectoryIterator($libPath) as $res) {
-                    if ($res->isDot() || substr($res->getFilename(), 0, 1) == '_') continue;
-                    $path = $res->getRealPath();
-                    if (!$res->isDir() || !\Tbx\Git::isGit($path)) continue;
-                    try {
-                        $v = \Tbx\Git::create($path, $input->getOption('dryRun'));
-                        $v->setInputOutput($input, $output);
-                        $this->writeInfo(ucwords($this->getName()) . ': ' . basename($v->getPath()));
-                        $tag = $v->getCurrentTag();
-                        if ($tag)
-                            $this->writeComment($tag);
-                            if ($input->getOption('nextTag')) {
-                                $nextTag = $v->lookupNewTag($input->getOptions());
-                                $this->writeBlue($nextTag);
-                            }
-                    } catch (\Exception $e) {
-                        $this->writeError($e->getMessage());
-                    }
-                }
+        $tagList = $vcs->getCurrentTags($input->getOptions());
+        foreach ($tagList as $name => $list) {
+            if ($input->getOption('noLibs') && $name != $projName) continue;
+            $nextTag = '';
+            if ($input->getOption('nextTag')) {
+                $nextTag = $list['next'];
             }
+            $this->getOutput()->writeln(sprintf($sformat, $name, $list['curr'], $nextTag));
         }
 
     }